Attended to review comments
