+++ integrations/bitbucket-cloud/.port/resources/port-app-config.yml
@@ -27,3 +27,4 @@ resources:
             defaultBranch: .mainbranch.name
           relations:
             project: '.project.uuid | gsub("[{-}]"; "")'
+
