+++ integrations/bitbucket-cloud/tests/entity_processors/test_file_entity_processor.py
@@ -1,7 +1,7 @@
 import json
 import pytest
 from unittest.mock import patch, AsyncMock
-from bitbucket_cloud.helpers.file_entity_handler import FileEntityProcessor
+from bitbucket_cloud.entity_processors.file_entity_processor import FileEntityProcessor
 
 MOCK_PORT_OCEAN_CONTEXT = AsyncMock()
 
@@ -13,14 +13,13 @@ async def test_file_entity_processor_search_json_success() -> None:
         "folder": {"path": "src", "commit": {"hash": "commit-hash"}},
     }
     pattern = "file://config.json"
-    expected_content = {"key": "value"}
-    json_str = json.dumps(expected_content)
+    expected_content = '{"key": "value"}'
 
     mock_client = AsyncMock()
-    mock_client.get_repository_files.return_value = json_str
+    mock_client.get_repository_files.return_value = expected_content
 
     with patch(
-        "bitbucket_cloud.helpers.file_entity_handler.init_client",
+        "bitbucket_cloud.entity_processors.file_entity_processor.init_client",
         return_value=mock_client,
     ):
         processor = FileEntityProcessor(
@@ -46,7 +45,7 @@ async def test_file_entity_processor_search_non_json_success() -> None:
     mock_client.get_repository_files.return_value = expected_content
 
     with patch(
-        "bitbucket_cloud.helpers.file_entity_handler.init_client",
+        "bitbucket_cloud.entity_processors.file_entity_processor.init_client",
         return_value=mock_client,
     ):
         processor = FileEntityProcessor(context=MOCK_PORT_OCEAN_CONTEXT)
@@ -70,7 +69,7 @@ async def test_file_entity_processor_search_error_handling() -> None:
     mock_client.get_repository_files.side_effect = Exception("Test exception")
 
     with patch(
-        "bitbucket_cloud.helpers.file_entity_handler.init_client",
+        "bitbucket_cloud.entity_processors.file_entity_processor.init_client",
         return_value=mock_client,
     ):
         processor = FileEntityProcessor(context=MOCK_PORT_OCEAN_CONTEXT)
@@ -107,13 +106,13 @@ async def test_file_entity_processor_search_missing_folder_commit() -> None:
     mock_client.get_repository_files.return_value = expected_content
 
     with patch(
-        "bitbucket_cloud.helpers.file_entity_handler.init_client",
+        "bitbucket_cloud.entity_processors.file_entity_processor.init_client",
         return_value=mock_client,
     ):
         processor = FileEntityProcessor(context=MOCK_PORT_OCEAN_CONTEXT)
         result = await processor._search(data, pattern)
         # In this case, since commit hash is missing, it should use the default branch ("main")
-        assert result == json.loads(expected_content)
+        assert result == expected_content
         mock_client.get_repository_files.assert_called_once_with(
             "test-repo", "main", "src/config.json"
         )
